Remove dubbo monitor module
