+++ .artifacts
@@ -62,10 +62,6 @@ dubbo-metrics-registry
 dubbo-metrics-config-center
 dubbo-metrics-netty
 dubbo-metrics-event
-dubbo-monitor
-dubbo-monitor-api
-dubbo-monitor-common
-dubbo-monitor-default
 dubbo-native
 dubbo-parent
 dubbo-plugin
