+++ dubbo-config/dubbo-config-api/pom.xml
@@ -80,11 +80,6 @@
 
     <dependency>
       <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-monitor-api</artifactId>
-      <version>${project.parent.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
       <artifactId>dubbo-remoting-api</artifactId>
       <version>${project.parent.version}</version>
     </dependency>
