+++ dubbo-distribution/dubbo-all/pom.xml
@@ -240,29 +240,6 @@
       <optional>true</optional>
     </dependency>
 
-    <!-- monitor -->
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-monitor-api</artifactId>
-      <version>${project.version}</version>
-      <scope>compile</scope>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-monitor-common</artifactId>
-      <version>${project.version}</version>
-      <scope>compile</scope>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-monitor-default</artifactId>
-      <version>${project.version}</version>
-      <scope>compile</scope>
-      <optional>true</optional>
-    </dependency>
-
     <!-- plugin -->
     <dependency>
       <groupId>org.apache.dubbo</groupId>
@@ -673,9 +650,6 @@
                   <include>org.apache.dubbo:dubbo-metrics-netty</include>
                   <include>org.apache.dubbo:dubbo-metrics-prometheus</include>
                   <include>org.apache.dubbo:dubbo-tracing</include>
-                  <include>org.apache.dubbo:dubbo-monitor-api</include>
-                  <include>org.apache.dubbo:dubbo-monitor-common</include>
-                  <include>org.apache.dubbo:dubbo-monitor-default</include>
                   <include>org.apache.dubbo:dubbo-qos</include>
                   <include>org.apache.dubbo:dubbo-qos-api</include>
                   <include>org.apache.dubbo:dubbo-security</include>
