+++ dubbo-metrics/dubbo-metrics-api/src/main/java/org/apache/dubbo/monitor/Constants.java
@@ -16,6 +16,7 @@
  */
 package org.apache.dubbo.monitor;
 
+@Deprecated
 public interface Constants {
     String DUBBO_PROVIDER = "dubbo.provider";
