+++ dubbo-metrics/dubbo-metrics-api/src/main/java/org/apache/dubbo/monitor/MonitorService.java
@@ -23,6 +23,7 @@ import java.util.List;
 /**
  * MonitorService. (SPI, Prototype, ThreadSafe)
  */
+@Deprecated
 public interface MonitorService {
 
     /**
