+++ dubbo-metrics/dubbo-metrics-default/src/main/java/org/apache/dubbo/monitor/support/MonitorFilter.java
@@ -61,6 +61,7 @@ import static org.apache.dubbo.rpc.Constants.OUTPUT_KEY;
 /**
  * MonitorFilter. (SPI, Singleton, ThreadSafe)
  */
+@Deprecated
 @Activate(group = {PROVIDER})
 public class MonitorFilter implements Filter, Filter.Listener {
