+++ dubbo-metrics/dubbo-metrics-default/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.cluster.filter.ClusterFilter
@@ -1,2 +1,3 @@
 observationsender=org.apache.dubbo.rpc.cluster.filter.support.ObservationSenderFilter
 metricsClusterFilter=org.apache.dubbo.rpc.cluster.filter.support.MetricsClusterFilter
+monitor=org.apache.dubbo.monitor.support.MonitorClusterFilter
