--- dubbo-monitor/dubbo-monitor-common/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.cluster.filter.ClusterFilter
@@ -1,1 +1,0 @@
-monitor=org.apache.dubbo.monitor.support.MonitorClusterFilter
