--- dubbo-monitor/dubbo-monitor-common/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.model.BuiltinServiceDetector
@@ -1,2 +1,0 @@
-monitor=org.apache.dubbo.monitor.support.MonitorServiceDetector
-metrics=org.apache.dubbo.monitor.support.MetricsServiceDetector
