--- dubbo-monitor/dubbo-monitor-default/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.monitor.MonitorFactory
@@ -1,1 +1,0 @@
-dubbo=org.apache.dubbo.monitor.dubbo.DubboMonitorFactory
