--- dubbo-monitor/dubbo-monitor-default/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.Filter
@@ -1,1 +1,0 @@
-metrics=org.apache.dubbo.monitor.dubbo.MetricsFilter
