+++ dubbo-test/dubbo-dependencies-all/pom.xml
@@ -213,22 +213,6 @@
       <version>${project.version}</version>
     </dependency>
 
-    <!-- monitor -->
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-monitor-api</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-monitor-common</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-monitor-default</artifactId>
-      <version>${project.version}</version>
-    </dependency>
     <!-- native -->
     <dependency>
       <groupId>org.apache.dubbo</groupId>
