+++ pom.xml
@@ -81,7 +81,6 @@
     <module>dubbo-rpc</module>
     <module>dubbo-cluster</module>
     <module>dubbo-registry</module>
-    <module>dubbo-monitor</module>
     <module>dubbo-config</module>
     <module>dubbo-demo</module>
     <module>dubbo-plugin</module>
